backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Segunda parte de prueba con LiteThinking"
authors = [
    {name = "Daniel Gonzalez",email = "dgonzalezca@unal.edu.co"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "google-genai (>=1.56.0,<2.0.0)",
    "google-auth (>=2.45.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "reportlab (>=4.4.7,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "markdown (>=3.10,<4.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "tzdata (>=2025.3,<2026.0)",
    "lite-dominio @ file:../dominio"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
lite-dominio = {develop = true}
